+++ processing/src/main/java/io/druid/query/aggregation/MinAggregatorFactory.java
@@ -1,181 +1,36 @@
 /*
- * Druid - a distributed column store.
- * Copyright 2012 - 2015 Metamarkets Group Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+* Licensed to Metamarkets Group Inc. (Metamarkets) under one
+* or more contributor license agreements. See the NOTICE file
+* distributed with this work for additional information
+* regarding copyright ownership. Metamarkets licenses this file
+* to you under the Apache License, Version 2.0 (the
+* "License"); you may not use this file except in compliance
+* with the License. You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing,
+* software distributed under the License is distributed on an
+* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+* KIND, either express or implied. See the License for the
+* specific language governing permissions and limitations
+* under the License.
+*/
 
 package io.druid.query.aggregation;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.google.common.base.Preconditions;
-import com.google.common.primitives.Doubles;
-import com.metamx.common.StringUtils;
-import io.druid.segment.ColumnSelectorFactory;
-
-import java.nio.ByteBuffer;
-import java.util.Arrays;
-import java.util.Comparator;
-import java.util.List;
 
 /**
- */
-public class MinAggregatorFactory implements AggregatorFactory
+ * This class can be removed when we dont need "min" aggregator backward compatibility
+  */
+public class MinAggregatorFactory extends DoubleMinAggregatorFactory
 {
-  private static final byte CACHE_TYPE_ID = 0x4;
-
-  private final String fieldName;
-  private final String name;
-
-  @JsonCreator
   public MinAggregatorFactory(
       @JsonProperty("name") String name,
-      @JsonProperty("fieldName") final String fieldName
+      @JsonProperty("fieldName") String fieldName
   )
   {
-    Preconditions.checkNotNull(name, "Must have a valid, non-null aggregator name");
-    Preconditions.checkNotNull(fieldName, "Must have a valid, non-null fieldName");
-
-    this.name = name;
-    this.fieldName = fieldName;
-  }
-
-  @Override
-  public Aggregator factorize(ColumnSelectorFactory metricFactory)
-  {
-    return new MinAggregator(name, metricFactory.makeFloatColumnSelector(fieldName));
-  }
-
-  @Override
-  public BufferAggregator factorizeBuffered(ColumnSelectorFactory metricFactory)
-  {
-    return new MinBufferAggregator(metricFactory.makeFloatColumnSelector(fieldName));
-  }
-
-  @Override
-  public Comparator getComparator()
-  {
-    return MinAggregator.COMPARATOR;
-  }
-
-  @Override
-  public Object combine(Object lhs, Object rhs)
-  {
-    return MinAggregator.combineValues(lhs, rhs);
-  }
-
-  @Override
-  public AggregatorFactory getCombiningFactory()
-  {
-    return new MinAggregatorFactory(name, name);
-  }
-
-  @Override
-  public List<AggregatorFactory> getRequiredColumns()
-  {
-    return Arrays.<AggregatorFactory>asList(new MinAggregatorFactory(fieldName, fieldName));
-  }
-
-  @Override
-  public Object deserialize(Object object)
-  {
-    // handle "NaN" / "Infinity" values serialized as strings in JSON
-    if (object instanceof String) {
-      return Double.parseDouble((String) object);
-    }
-    return object;
-  }
-
-  @Override
-  public Object finalizeComputation(Object object)
-  {
-    return object;
-  }
-
-  @JsonProperty
-  public String getFieldName()
-  {
-    return fieldName;
-  }
-
-  @Override
-  @JsonProperty
-  public String getName()
-  {
-    return name;
-  }
-
-  @Override
-  public List<String> requiredFields()
-  {
-    return Arrays.asList(fieldName);
-  }
-
-  @Override
-  public byte[] getCacheKey()
-  {
-    byte[] fieldNameBytes = StringUtils.toUtf8(fieldName);
-
-    return ByteBuffer.allocate(1 + fieldNameBytes.length).put(CACHE_TYPE_ID).put(fieldNameBytes).array();
-  }
-
-  @Override
-  public String getTypeName()
-  {
-    return "float";
-  }
-
-  @Override
-  public int getMaxIntermediateSize()
-  {
-    return Doubles.BYTES;
-  }
-
-  @Override
-  public Object getAggregatorStartValue()
-  {
-    return Double.POSITIVE_INFINITY;
-  }
-
-  @Override
-  public String toString()
-  {
-    return "MinAggregatorFactory{" +
-           "fieldName='" + fieldName + '\'' +
-           ", name='" + name + '\'' +
-           '}';
-  }
-
-  @Override
-  public boolean equals(Object o)
-  {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-
-    MinAggregatorFactory that = (MinAggregatorFactory) o;
-
-    if (fieldName != null ? !fieldName.equals(that.fieldName) : that.fieldName != null) return false;
-    if (name != null ? !name.equals(that.name) : that.name != null) return false;
-
-    return true;
-  }
-
-  @Override
-  public int hashCode()
-  {
-    int result = fieldName != null ? fieldName.hashCode() : 0;
-    result = 31 * result + (name != null ? name.hashCode() : 0);
-    return result;
+    super(name, fieldName);
   }
 }
