+++ processing/src/test/java/io/druid/segment/AppendTest.java
@@ -28,9 +28,9 @@ import io.druid.query.Result;
 import io.druid.query.TestQueryRunners;
 import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.aggregation.CountAggregatorFactory;
+import io.druid.query.aggregation.DoubleMaxAggregatorFactory;
 import io.druid.query.aggregation.DoubleSumAggregatorFactory;
-import io.druid.query.aggregation.MaxAggregatorFactory;
-import io.druid.query.aggregation.MinAggregatorFactory;
+import io.druid.query.aggregation.DoubleMinAggregatorFactory;
 import io.druid.query.aggregation.PostAggregator;
 import io.druid.query.aggregation.hyperloglog.HyperUniquesAggregatorFactory;
 import io.druid.query.aggregation.post.ArithmeticPostAggregator;
@@ -563,8 +563,8 @@ public class AppendTest
                                           Iterables.concat(
                                               commonAggregators,
                                               Lists.newArrayList(
-                                                  new MaxAggregatorFactory("maxIndex", "index"),
-                                                  new MinAggregatorFactory("minIndex", "index")
+                                                  new DoubleMaxAggregatorFactory("maxIndex", "index"),
+                                                  new DoubleMinAggregatorFactory("minIndex", "index")
                                               )
                                           )
                                       )
@@ -587,8 +587,8 @@ public class AppendTest
                          Iterables.concat(
                              commonAggregators,
                              Lists.newArrayList(
-                                 new MaxAggregatorFactory("maxIndex", "index"),
-                                 new MinAggregatorFactory("minIndex", "index")
+                                 new DoubleMaxAggregatorFactory("maxIndex", "index"),
+                                 new DoubleMinAggregatorFactory("minIndex", "index")
                              )
                          )
                      )
@@ -623,8 +623,8 @@ public class AppendTest
                          Iterables.concat(
                              commonAggregators,
                              Lists.newArrayList(
-                                 new MaxAggregatorFactory("maxIndex", "index"),
-                                 new MinAggregatorFactory("minIndex", "index")
+                                 new DoubleMaxAggregatorFactory("maxIndex", "index"),
+                                 new DoubleMinAggregatorFactory("minIndex", "index")
                              )
                          )
                      )
@@ -647,8 +647,8 @@ public class AppendTest
                 Iterables.concat(
                     commonAggregators,
                     Lists.newArrayList(
-                        new MaxAggregatorFactory("maxIndex", "index"),
-                        new MinAggregatorFactory("minIndex", "index")
+                        new DoubleMaxAggregatorFactory("maxIndex", "index"),
+                        new DoubleMinAggregatorFactory("minIndex", "index")
                     )
                 )
             )
@@ -686,8 +686,8 @@ public class AppendTest
                 Iterables.concat(
                     commonAggregators,
                     Lists.newArrayList(
-                        new MaxAggregatorFactory("maxIndex", "index"),
-                        new MinAggregatorFactory("minIndex", "index")
+                        new DoubleMaxAggregatorFactory("maxIndex", "index"),
+                        new DoubleMinAggregatorFactory("minIndex", "index")
                     )
                 )
             )
